backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
paho-mqtt==1.6.1